+++ include/jsoncons_ext/jsonpath/token_evaluator.hpp
@@ -2420,12 +2420,12 @@ namespace detail {
 
         struct MyHash
         {
-            std::uintmax_t operator()(string_type const& s) const noexcept
+            std::size_t operator()(string_type const& s) const noexcept
             {
                 const int p = 31;
                 const int m = static_cast<int>(1e9) + 9;
-                std::uintmax_t hash_value = 0;
-                std::uintmax_t p_pow = 1;
+                std::size_t hash_value = 0;
+                std::size_t p_pow = 1;
                 for (char_type c : s) {
                     hash_value = (hash_value + (c - 'a' + 1) * p_pow) % m;
                     p_pow = (p_pow * p) % m;
